#2552 Update for CSLA 6
