+++ Samples/CustomActivator/CustomActivator/ITestItem.cs
@@ -1,0 +1,9 @@
+﻿using Csla;
+
+namespace CustomActivator
+{
+  public interface ITestItem : IBusinessBase
+  {
+    string Name { get; set; }
+  }
+}
