+++ Samples/CustomActivator/CustomActivator/packages.config
@@ -1,15 +1,27 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Csla" version="5.0.0" targetFramework="net461" />
-  <package id="Microsoft.Extensions.Configuration" version="3.0.0" targetFramework="net461" />
-  <package id="Microsoft.Extensions.Configuration.Abstractions" version="3.0.0" targetFramework="net461" />
-  <package id="Microsoft.Extensions.Configuration.Binder" version="3.0.0" targetFramework="net461" />
-  <package id="Microsoft.Extensions.DependencyInjection" version="3.0.0" targetFramework="net461" />
-  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="3.0.0" targetFramework="net461" />
-  <package id="Microsoft.Extensions.Primitives" version="3.0.0" targetFramework="net461" />
-  <package id="System.Buffers" version="4.5.0" targetFramework="net461" />
-  <package id="System.ComponentModel.Annotations" version="4.6.0" targetFramework="net461" />
-  <package id="System.Memory" version="4.5.3" targetFramework="net461" />
+  <package id="Csla" version="6.0.0-R22040302" targetFramework="net48" />
+  <package id="Microsoft.Bcl.AsyncInterfaces" version="6.0.0" targetFramework="net48" />
+  <package id="Microsoft.Extensions.Configuration" version="6.0.1" targetFramework="net48" />
+  <package id="Microsoft.Extensions.Configuration.Abstractions" version="6.0.0" targetFramework="net48" />
+  <package id="Microsoft.Extensions.Configuration.Binder" version="6.0.0" targetFramework="net48" />
+  <package id="Microsoft.Extensions.DependencyInjection" version="6.0.0" targetFramework="net48" />
+  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="6.0.0" targetFramework="net48" />
+  <package id="Microsoft.Extensions.Primitives" version="6.0.0" targetFramework="net48" />
+  <package id="System.Buffers" version="4.5.1" targetFramework="net48" />
+  <package id="System.ComponentModel" version="4.3.0" targetFramework="net48" />
+  <package id="System.ComponentModel.Annotations" version="5.0.0" targetFramework="net48" />
+  <package id="System.ComponentModel.Primitives" version="4.3.0" targetFramework="net48" />
+  <package id="System.Data.Common" version="4.3.0" targetFramework="net48" />
+  <package id="System.Memory" version="4.5.4" targetFramework="net48" />
   <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net461" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.6.0" targetFramework="net461" />
+  <package id="System.Reflection" version="4.3.0" targetFramework="net48" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="6.0.0" targetFramework="net48" />
+  <package id="System.Runtime.Serialization.Formatters" version="4.3.0" targetFramework="net48" />
+  <package id="System.Runtime.Serialization.Primitives" version="4.3.0" targetFramework="net48" />
+  <package id="System.Runtime.Serialization.Xml" version="4.3.0" targetFramework="net48" />
+  <package id="System.Security.Principal" version="4.3.0" targetFramework="net48" />
+  <package id="System.Threading.Tasks.Extensions" version="4.5.4" targetFramework="net461" />
+  <package id="System.Threading.ThreadPool" version="4.3.0" targetFramework="net48" />
+  <package id="System.ValueTuple" version="4.5.0" targetFramework="net48" />
 </packages>
